Crate liva_parser

The parser for the liva source code

Example

use std::error::Error;

use liva_parser::print_ast;

fn main() -> Result<(), Box<dyn Error>> {
    print_ast("lib.lv");
    Ok(())
}

Modules

expression

Logic for expressions and their representation

literals

Collection of all possible literals

statement

Collection of all statements

tokens

Collection of all parsers and structs which represent tokens, operators and keywords

Functions

parse_source

Entry point to the parser, which parse the liva language

print_ast

Prints the AST of the parsed liva source code

Type Definitions

Res